requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
xgboost==1.7.3
pyyaml==6.0
joblib==1.2.0